+++ sdk/src/main/java/io/opentdf/platform/sdk/Config.java
@@ -10,6 +10,10 @@ import io.opentdf.platform.policy.Value;
 import java.util.*;
 import java.util.function.Consumer;
 
+/**
+ * Configuration class for setting various configurations related to TDF.
+ * Contains nested classes and enums for specific configuration settings.
+ */
 public class Config {
 
     public static final int TDF3_KEY_SIZE = 2048;
@@ -87,12 +91,14 @@ public class Config {
 
     public static class TDFReaderConfig {
         // Optional Map of Assertion Verification Keys
-        AssertionVerificationKeys assertionVerificationKeys;
+        AssertionVerificationKeys assertionVerificationKeys = new AssertionVerificationKeys();
+        boolean disableAssertionVerification;
     }
 
     @SafeVarargs
     public static TDFReaderConfig newTDFReaderConfig(Consumer<TDFReaderConfig>... options) {
         TDFReaderConfig config = new TDFReaderConfig();
+        config.disableAssertionVerification = false;
         for (Consumer<TDFReaderConfig> option : options) {
             option.accept(config);
         }
@@ -102,6 +108,10 @@ public class Config {
     public static Consumer<TDFReaderConfig> withAssertionVerificationKeys(
             AssertionVerificationKeys assertionVerificationKeys) {
         return (TDFReaderConfig config) -> config.assertionVerificationKeys = assertionVerificationKeys;
+    }
+
+    public static Consumer<TDFReaderConfig> withDisableAssertionVerification(boolean disable) {
+        return (TDFReaderConfig config) -> config.disableAssertionVerification = disable;
     }
 
     public static class TDFConfig {