+++ sdk/src/main/java/io/opentdf/platform/sdk/Manifest.java
@@ -10,6 +10,9 @@ import com.nimbusds.jose.crypto.RSASSASigner;
 import com.nimbusds.jose.crypto.RSASSAVerifier;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
+
+import io.opentdf.platform.sdk.TDF.AssertionException;
+
 import org.apache.commons.codec.binary.Hex;
 import org.erdtman.jcs.JsonCanonicalizer;
 
@@ -27,6 +30,10 @@ import java.util.Base64;
 import java.util.List;
 import java.util.Objects;
 
+/**
+ * The Manifest class represents a detailed structure encapsulating various aspects
+ * of data integrity, encryption, payload, and assertions within a certain context.
+ */
 public class Manifest {
 
     private static final String kAssertionHash = "assertionHash";
@@ -377,7 +384,7 @@ public class Manifest {
         public Assertion.HashValues verify(AssertionConfig.AssertionKey assertionKey)
                 throws ParseException, JOSEException {
             if (binding == null) {
-                throw new SDKException("Binding is null in assertion");
+                throw new AssertionException("Binding is null in assertion", this.id);
             }
 
             String signatureString = binding.signature;